+++ crates/ec_cli/src/cli/debug/mod.rs
@@ -456,7 +456,10 @@ impl DebugSubcommand {
                 }
                 println!("\n\n\n-------\nFinished writing to {}", outfile.display());
                 println!("Please send this file to the development team");
-                println!("Or attach it to a Github issue (run '{}')", "fig issue".magenta());
+                println!(
+                    "Or attach it to a Github issue (run '{}')",
+                    format!("{CLI_BINARY_NAME} issue").magenta()
+                );
             },
             #[cfg(target_os = "macos")]
             DebugSubcommand::VerifyCodesign => {